djangosite/requirements.txt
Django==1.11.5
django-bootstrap3==9.0.0
djangorestframework==3.6.4
olefile==0.44
Pillow==4.2.1
pytz==2017.2
reportlab==3.4.0
